Update ReceivePaymentRequest expiry type

+++ libs/sdk-react-native/android/src/main/java/com/breezsdk/BreezSDKMapper.kt
@@ -110,7 +110,7 @@ fun asReceivePaymentRequest(reqData: ReadableMap): ReceivePaymentRequest? {
         val preimage = reqData.getArray("preimage")?.let { asUByteList(it) }
         val openingFeeParams = reqData.getMap("openingFeeParams")?.let{ asOpeningFeeParams(it) }
         val useDescriptionHash = if (hasNonNullKey(reqData, "useDescriptionHash")) reqData.getBoolean("useDescriptionHash") else null
-        val expiry = if (hasNonNullKey(reqData, "expiry")) reqData.getDouble("expiry").toULong() else null
+        val expiry = if (hasNonNullKey(reqData, "expiry")) reqData.getInt("expiry").toUInt() else null
         val cltv = if (hasNonNullKey(reqData, "cltv")) reqData.getInt("cltv").toUInt() else null
         return ReceivePaymentRequest(amountSats.toULong(), description!!, preimage, openingFeeParams, useDescriptionHash, expiry, cltv)
     }
